+++ common/src/main/java/org/infernalstudios/questlog/core/quests/objectives/Objective.java
@@ -17,7 +17,7 @@ public abstract class Objective implements NbtSaveable, WithDisplayData<Objectiv
     private final int requiredAmount;
     private Quest parent;
     private int units;
-    private boolean isPartOfRequirements = false;
+    private boolean isPartOfPrerequisites = false;
 
     public Objective(JsonObject definition) {
         this.requiredAmount = JsonUtils.getOrDefault(definition, "required_amount", 1);
@@ -30,8 +30,13 @@ public abstract class Objective implements NbtSaveable, WithDisplayData<Objectiv
     public void registerEventListeners() {
     }
 
+    public void markAsPrerequisite() {
+        this.isPartOfPrerequisites = true;
+    }
+
+    @Deprecated
     public void markAsRequirement() {
-        this.isPartOfRequirements = true;
+        this.markAsPrerequisite();
     }
 
     @Nullable
@@ -48,7 +53,7 @@ public abstract class Objective implements NbtSaveable, WithDisplayData<Objectiv
     }
 
     public void setUnits(int units) {
-        if (this.getParent() != null && !this.getParent().isTriggered() && !this.isPartOfRequirements) {
+        if (this.getParent() != null && !this.getParent().isTriggered() && !this.isPartOfPrerequisites) {
             return;
         }